I updated the paper.
I do not use Word much, and I do not know if you use Zotero or Endnote, so I just created the additional_bibliography.bib file for the randomForest extensions.

I hope that I answered Zoli's questions correctly and fully.

There was a major update of the dplyr package last week, and I am still tweaking the create_choropleth() function for maps in the EU supplement material.  It works almost as earlier, but I want to improve aesthetically the way it handles missing cases.

I will tweak this and upload the dataset and its PDF long-form documentation to figshare separately.  But I do not think that any changes in this (which are cosmetic) affect the review process.

The new doi's are not yet added to the comments in the helper files, but I will do these later today.